+++ src/utils/index.ts
@@ -2,3 +2,10 @@ export { goToSection } from './goToSection.util'
 export { openInNewTab } from './openInNewTab.util'
 export { getYearsOfExperience } from './getYearsOfExperience.util'
 export { formatPlusSignByLang } from './formatPlusSignByLang.util'
+export {
+  formatPeriod,
+  formatPeriodWithDuration,
+  calculateDuration,
+  formatDuration,
+  formatDate,
+} from './experiencePeriod.util'